added more est and made dtype consitent for obs

+++ gym_bb/tasks/monopod_v1_balancing_fixed_hip_and_boom_yaw.py
@@ -41,7 +41,7 @@ class MonopodV1BalancingFixedHipAndBoomYaw(MonopodBase):
         # Create the action space
         action_space = gym.spaces.Box(low=np.array([-self.max_torque_upper_leg, -self.max_torque_lower_leg]),
                                       high=np.array([self.max_torque_upper_leg,  self.max_torque_lower_leg]),
-                                      dtype=np.float32)
+                                      dtype=np.float64)
         # Configure reset limits
         high = np.array([
             self._u_limit,
@@ -63,14 +63,14 @@ class MonopodV1BalancingFixedHipAndBoomYaw(MonopodBase):
         # Configure the reset space - this is used to check if it exists inside the reset space when deciding whether to reset.
         self.reset_space = gym.spaces.Box(low=low,
                                           high=high,
-                                          dtype=np.float32)
+                                          dtype=np.float64)
 
         # Configure the observation space
         obs_high = high.copy() * 1.2
         obs_low = low.copy() * 1.2
         observation_space = gym.spaces.Box(low=obs_low,
                                            high=obs_high,
-                                           dtype=np.float32)
+                                           dtype=np.float64)
 
         return action_space, observation_space
